+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Change Log
 
+### [Unreleased][unreleased]
+
+#### Added
+- timeout to redis connect
+
 ### [0.1.0] - 2016/04/17
 - First tagged version
 - possible drivers: \Redis, \Predis\Client